Issue #65: Reorganized large modules in urls and views (billing module, root urls) into specific submodules.
