+++ dds_registration/urls/membership.py
@@ -1,0 +1,27 @@
+# @module XXX
+# @changed 2024.04.02, 00:28
+
+from django.urls import path
+
+from ..views import membership as membership_views
+
+
+urlpatterns = [
+    # Membership (TODO: Some routes will be moved to `billing`)...
+    path(
+        "membership/start",
+        membership_views.membership_start,
+        name="membership_start",
+    ),
+    # path(
+    #     # UNUSED?
+    #     "membership/proceed/<str:membership_type>",
+    #     membership_views.membership_proceed,
+    #     name="membership_proceed",
+    # ),
+    path(
+        "membership/proceed/success",
+        membership_views.membership_proceed_success,
+        name="membership_proceed_success",
+    ),
+]
